Revert "use __construct() for constructor"
This reverts commit 26ad947c5d9541a44721b9d972441fe708f41860.

+++ zp-core/zp-extensions/galleryArticles.php
@@ -41,10 +41,10 @@ class galleryArticles {
 			setOptionDefault('galleryArticles_image_text', getAllTranslations($text));
 			setOptionDefault('galleryArticles_size', 80);
 			setOptionDefault('galleryArticles_protected', 0);
-
-			require_once(SERVERPATH . '/' . ZENFOLDER . '/' . PLUGIN_FOLDER . '/cacheManager.php');
-			cacheManager::deleteThemeCacheSizes('galleryArticles');
-			cacheManager::addThemeCacheSize('galleryArticles', getOption('galleryArticles_size'), NULL, NULL, NULL, NULL, NULL, NULL, false, getOption('fullimage_watermark'), NULL, NULL);
+			if (class_exists('cacheManager')) {
+				cacheManager::deleteThemeCacheSizes('galleryArticles');
+				cacheManager::addThemeCacheSize('galleryArticles', getOption('galleryArticles_size'), NULL, NULL, NULL, NULL, NULL, NULL, false, getOption('fullimage_watermark'), NULL, NULL);
+			}
 		}
 	}
